requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
fpdf2>=2.7.5
Pillow>=10.0.0
openpyxl>=3.0.0
